データロードしてからレンダリングするよう修正
メインページを子コンポーネントに分離し、ロードしたデータを子に渡すことで
Unknown gift idエラーが発生する不具合を修正。

また不要なファイルを削除。
