--- app/components/CSR.tsx
@@ -1,5 +1,0 @@
-import dynamic from "next/dynamic";
-
-const CSR = dynamic(() => import("./CSRInner"), { ssr: false });
-
-export default CSR;
