--- app/components/CSRInner.tsx
@@ -1,7 +1,0 @@
-import { JSX } from "react";
-
-const CSRInner = ({ children }: { children: JSX.Element | JSX.Element[] }) => {
-  return <>{children}</>;
-};
-
-export default CSRInner;
